Service Web fonctionnel
Service web fonctionnel avec exemples.

A faire :
 - Revoir "PERSIST" et "MERGE"
 - Faire l'interface de gestion des données

+++ src/main/java/univ/tours/WeBuy/dao/AmiJpaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import univ.tours.WeBuy.entities.Ami;
 
 @RepositoryRestResource
-public interface AmiJpaRepository extends JpaRepository<Ami,Long>{
+public interface AmiJpaRepository extends JpaRepository<Ami,Integer>{
 
 }
