+++ src/main/java/univ/tours/WeBuy/dao/MagasinJpaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import univ.tours.WeBuy.entities.Magasin;
 
 @RepositoryRestResource
-public interface MagasinJpaRepository extends JpaRepository<Magasin,Long> {
+public interface MagasinJpaRepository extends JpaRepository<Magasin,Integer> {
 
 }
