+++ src/main/java/univ/tours/WeBuy/dao/ProduitJpaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import univ.tours.WeBuy.entities.Produit;
 
 @RepositoryRestResource
-public interface ProduitJpaRepository extends JpaRepository<Produit,Long> {
+public interface ProduitJpaRepository extends JpaRepository<Produit,Integer> {
 
 }
