+++ src/main/java/univ/tours/WeBuy/dao/UtilisateurJpaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import univ.tours.WeBuy.entities.Utilisateur;
 
 @RepositoryRestResource
-public interface UtilisateurJpaRepository extends JpaRepository<Utilisateur,Long> {
+public interface UtilisateurJpaRepository extends JpaRepository<Utilisateur,Integer> {
 
 }
